+++ CHANGELOG.md
@@ -1,7 +1,16 @@
 
-# Changelog for OHLC v1.1.8
+# Change log
 
-Fixed bug where new candle generation was created with wrong year number.
+## v1.2.0 - 26.06.2022
+
+### Added
+- Factory module **OHLCFactory** which can be used to generate data.
+- **gen_candles/3** function to generate OHLC candles.
+
+### Deprecated
+- **OHLCHelper.gen_trades/6** use **OHLCFactory.gen_trades/1 instead**
+- **OHLCHelper.generate_empty_candle/1** use **OHLCFactory.gen_empty_candle/1 instead**
+